requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.1.4
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
six==1.11.0
Werkzeug==0.14.1
